+++ src/comp/c_udp.c
@@ -444,9 +444,9 @@ bad_context:
  */
 int c_udp_encode(struct c_context *const context,
                  const struct ip_packet *ip,
-                 const int packet_size,
+                 const size_t packet_size,
                  unsigned char *const dest,
-                 const int dest_size,
+                 const size_t dest_size,
                  rohc_packet_t *const packet_type,
                  int *const payload_offset)
 {
