+++ src/comp/rohc_comp_internals.h
@@ -266,9 +266,9 @@ struct c_profile
 	 */
 	int (*encode)(struct c_context *const context,
 	              const struct ip_packet *packet,
-	              const int packet_size,
+	              const size_t packet_size,
 	              unsigned char *const dest,
-	              const int dest_size,
+	              const size_t dest_size,
 	              rohc_packet_t *const packet_type,
 	              int *const payload_offset);
